bc/requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
idna==3.4
jdatetime==4.1.0
PyYAML==6.0
requests==2.28.1
urllib3==1.26.12
